pyproject.toml
[project]
name = "cli-podcast-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "requests>=2.32.5",
]
